Rename to `unlink-image-auto-links`
