+++ readme.md
@@ -354,7 +354,7 @@ Thanks for contributing! 🦋🙌
 - [](# "same-page-definition-jump") [Avoids re-loading the page when jumping to function definition in the current file.](https://user-images.githubusercontent.com/16872793/90833649-7a5e2f80-e316-11ea-827d-a4e3ac8ced69.png)
 - [](# "linkify-full-profile-readme-title") [Linkifies the readme text on profile pages.](https://user-images.githubusercontent.com/16872793/90910173-ebe4bf00-e3a4-11ea-8fc5-aea3d1a2e5e5.png)
 - [](# "convert-pr-to-draft-improvements") [Moves the "Convert PR to Draft" button to the mergeability box and adds visual feedback to its confirm button.](https://user-images.githubusercontent.com/1402241/95644892-885f3f80-0a7f-11eb-8428-8e0fb0c8dfa5.gif)
-- [](# "clean-image-auto-links") [Removes unnecessary image auto links if they point to themselves.](https://user-images.githubusercontent.com/44045911/112277464-bb5b5f80-8cbc-11eb-9dc0-0268fc3a2b8a.png)
+- [](# "unlink-image-auto-links") [Removes unnecessary image auto links if they point to themselves.](https://user-images.githubusercontent.com/44045911/112277464-bb5b5f80-8cbc-11eb-9dc0-0268fc3a2b8a.png)
 
 <!-- Refer to style guide above. Keep this message between sections. -->
